+++ Dia01-github/semana08/dia03/01-js-fetch-crud/public/index.html
@@ -60,16 +60,19 @@
           <input type="date" name="estreno" placeholder="Estreno" required>
           <select name="genero" required>
           <option value="">Seleccione una Opción</option>
-          <option value="1">Comedia></option>
-          <option value="1">Comedia></option>
-          <option value="1">Comedia></option>
-          <option value="1">Comedia></option>
+          <option value="1">Comedia</option>
+          <option value="1">Acción</option>
+          <option value="1">Aminada</option>
+          <option value="1">Aventura</option>
         </select>
+        <textarea name="resumen" placeholder="Resumen de mi película"></textarea>
+
+        <input type="submit" value="Guardar" />
         </form>
       </div>
     </div>
   </section>
 
-  <script src="./main.js"></script>
+  <script src="./main.js" type="module"></script>
 </body>
 </html>
