+++ Dia01-github/semana08/dia03/01-js-fetch-crud/public/services.js
@@ -1,0 +1,7 @@
+export const fetchPeliculas = async () => {
+    const url = 'http://localhost:3000/peliculas'
+  
+    const response = await fetch(url)
+  
+    return await response.json()
+  }
